Applied #28 TimeTagArgument doesn't hold time
The docstring for TimeTagArgument specifies that a time tag is the seconds since midnight on 1/1/1900. From then to the UNIX Epoch (1/1/1970) is 2208988800 seconds; from the Epoch to about now is 1297985487.378675 seconds. So from 1/1/1900 to now is 3506974287.378675 seconds. However:

>>> txosc.osc.TimeTagArgument(3506974287.378675).toBinary()
__main__:1: DeprecationWarning: 'l' format requires -2147483648 <= number <= 2147483647
"\xd1\x08.O\x16\x92\x1f'"

and

>>> txosc.osc.TimeTagArgument.fromBinary(txosc.osc.TimeTagArgument(3506974287.378675).toBinary())[0].value
-787993008.62132502

which is clearly not 3506974287.378675.

I think this can be fixed by replacing the three occurrences of struct.pack('>ll'... with struct.pack('>qq'... in osc.py.